requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
python-decouple==3.8
openpyxl==3.1.2
pandas==2.1.4
qrcode==7.4.2
line-bot-sdk==3.5.0
requests==2.31.0
